// Frontend/sistemaTickets/src/app/Model/ITicketSupport.ts
export interface ITicketSupport {
  idControl: number | null;
  consecutive: number | null;
  date: string;
  title: string;
  aerea: string;
  description: string;
  status: string;
  priority: string;
  photoDescription: string;
  softwareApplication: number;
  codeConnectionSoftware: number;
  enabled: boolean
}